+++ btb/__init__.py
@@ -4,4 +4,4 @@
 
 __author__ = """MIT Data To AI Lab"""
 __email__ = 'dailabmit@gmail.com'
-__version__ = '0.3.3'
+__version__ = '0.3.4.dev0'